requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-dotenv
passlib[bcrypt]
pytest
httpx
slowapi
